simplify implementation #100 leave builder pattern for future iterations
